// src/data/rating/types/GetDetailRatingsType.ts
export type GetDetailRatingsDataResponse = {
	uuid: string;
	food_rating_uuid: string;
	product_uuid: string;
	product_name: string;
	addon_information: Array<AddonInformation>;
	rating: number;
	review: Array<string>;
};

type AddonInformation = {
	addon_name: string;
	addon_price: number;
	addon_variants: Array<AddonVariant>;
};

type AddonVariant = {
	variant_name: string;
	variant_price: number;
};